update chain params

+++ chaincfg/params.go
@@ -235,7 +235,7 @@ type Params struct {
 	MaxCoinbaseScriptLen int
 	// MedianTimeBlocks is the number of previous blocks which should be
 	// used to calculate the median time used to validate block timestamps.
-	MedianTimeBlocks int64
+	MedianTimeBlocks int
 	// SerializedHeightVersion is the block version which changed block
 	// coinbases to start with the serialized block height.
 	SerializedHeightVersion int32
